% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_methods.R
\name{calc_gene_stats}
\alias{calc_gene_stats}
\title{Method to calculate gene average expression and variance for an expression matrix}
\usage{
calc_gene_stats(dat, hgnc_data_name, hgnc_data_file, hgnc_version, hgnc_taiga)
}
\arguments{
\item{dat:}{data object containing tumor and cell line expression data and annotations produced by running load_data}

\item{hgnc_data_name:}{if hgnc_taiga = TRUE, then the data.name of the taiga file containing the HGNC gene annotations,
if hgnc_taiga=FALSE, then the file path to the local folder containing the HGNC gene annotations}

\item{hgnc_data_file:}{if hgnc_taiga = TRUE, then the data.file of the taiga file containing the HGNC gene annotations,
if hgnc_taiga=FALSE, then the name of the file the HGNC gne annotations}

\item{hgnc_version:}{parameter to specify the version to pull from taiga for the HGNC gene annotations, default set to NULL}

\item{hgnc_taiga:}{if TRUE then pulls the HGNC gene annotations from taiga, if FALSE then finds HGNC gene annotations in local folder}
}
\value{
gene stats matrix
}
\description{
calculate the average gene expression and variance
}
