% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_methods.R
\name{calc_tumor_CL_cor}
\alias{calc_tumor_CL_cor}
\title{Method to calculate the correlation between cell lines and tumor in the Celligner aligned data}
\usage{
calc_tumor_CL_cor(Celligner_aligned_data, Celligner_info)
}
\arguments{
\item{Celligner_aligned_data:}{Celligner aligned data matrix of samples (cells line and tumors) by genes}

\item{Celligner_info:}{annotation file of cell line and tumor samples with a column 'type' marking samples as either
cell lines or tumors and a column 'sampleID' that matches the row names of Celligner_aligned_data}
}
\value{
matrix of correlations that is tumors by cell lines
}
\description{
calculate the correlation between cell line and tumor samples in the Celligner aligned data
}
