% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_helpers.R
\name{check_NAs}
\alias{check_NAs}
\title{check for NAs in the expression data and remove samples with NAs}
\usage{
check_NAs(mat)
}
\arguments{
\item{mat:}{matrix of gene expression data that is samples by genes}
}
\value{
matrix of gene expression data, removing samples that have NAs
}
\description{
check for NAs in the expression data and remove samples with NAs
}
