% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_methods.R
\name{cluster_data}
\alias{cluster_data}
\title{Method to take in a Seurat object and run default Seurat clustering algorithm}
\usage{
cluster_data(seu_obj)
}
\arguments{
\item{seu_obj:}{seurat object containing expression data and sample annotations.
Expects PCA for the seurat object has already been calculated.}
}
\value{
Seurat object with cluster annotations
}
\description{
cluster data in seurat object, using default Seurat clustering method. Clsuters data
within PCA space using the number of dimensions provided in global$n_PC_dims (default is 70)
}
