% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_helpers.R
\name{.average_correction}
\alias{.average_correction}
\title{calculate the average correction vector}
\usage{
.average_correction(refdata, mnn1, curdata, mnn2)
}
\arguments{
\item{refdata:}{matrix of samples by genes of cPC corrected data that serves as the reference data in the MNN alignment.
In the standard Celligner pipeline this the cell line data.}

\item{mnn1:}{mnn1 pairs}

\item{curdata:}{matrix of samples by genes of cPC corrected data that is corrected in the MNN alignment and projected onto the reference data.
In the standard Celligner pipeline this the tumor data.}

\item{mnn2:}{mnn2 pairs}
}
\value{
correction vector and pairs
}
\description{
Computes correction vectors for each MNN pair, and then averages them for each MNN-involved cell in the second batch.
Copied from dev version of scran (2018-10-28), with slight modifications as noted https://github.com/MarioniLab/scran
}
