% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_helpers.R
\name{.center_along_batch_vector}
\alias{.center_along_batch_vector}
\title{centers samples within each batch}
\usage{
.center_along_batch_vector(mat, batch.vec)
}
\arguments{
\item{mat:}{matrix of samples by genes}

\item{batch.vec:}{batch vector}
}
\value{
correction vector and pairs
}
\description{
Projecting along the batch vector, and shifting all samples to the center within each batch.
This removes any variation along the overall batch vector within each matrix.
}
