% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_helpers.R
\name{.compute_tricube_average}
\alias{.compute_tricube_average}
\title{compute tricube averages}
\usage{
.compute_tricube_average(vals, indices, distances, bandwidth = NULL, ndist = 3)
}
\arguments{
\item{values:}{correction vector}

\item{indices:}{nxk matrix for the nearest neighbor indice}

\item{distances:}{nxk matrix for the nearest neighbor Euclidea distances}

\item{bandwidth:}{Is set at 'ndist' times the median distance, if not specified.}

\item{ndist:}{By default is 3.}
}
\description{
Centralized function to compute tricube averages.
}
