% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_helpers.R
\name{.tricube_weighted_correction}
\alias{.tricube_weighted_correction}
\title{tricube-weighted correction}
\usage{
.tricube_weighted_correction(
  curdata,
  correction,
  in.mnn,
  k = 20,
  ndist = 3,
  subset_genes,
  BNPARAM = NULL,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{curdata:}{target matrix of samples by genes}

\item{correction:}{corrected vector}

\item{in.mnn:}{mnn pairs}

\item{k:}{k values, default 20}

\item{ndist:}{A numeric scalar specifying the threshold beyond which neighbors are to be ignored when computing correction vectors.
By default is 3.}

\item{subset_genes:}{genes used to identify mutual nearest neighbors}

\item{BNPARAM:}{default NULL}

\item{BPPARAM:}{default BiocParallel::SerialParam()}
}
\value{
MNN corrected data
}
\description{
Computing tricube-weighted correction vectors for individual samples,
using the nearest neighbouring samples involved in MNN pairs.
Modified to use FNN rather than queryKNN for nearest neighbor finding
}
