% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_methods.R
\name{find_differentially_expressed_genes}
\alias{find_differentially_expressed_genes}
\title{Method to find genes that are differentially expressed between clusters within the expression data}
\usage{
find_differentially_expressed_genes(seu_obj)
}
\arguments{
\item{seu_obj:}{seurat object containing expression data and sample annotations. Expects data in the Seurat object
slot scale.data and a column 'seurat_clusters' within the meta.data of the Seurat object.}
}
\value{
table with gene level stats
}
\description{
find genes that are differentially expressed between clusters within the expression data
}
