% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_helpers.R
\name{get_cluster_averages}
\alias{get_cluster_averages}
\title{calculate the average expression per cluster}
\usage{
get_cluster_averages(mat, cluster_df)
}
\arguments{
\item{mat:}{sample by genes matrix of expression data}

\item{cluster_df:}{table of sample metadata that includes a column 'seurat_clusters',
containing transcriptional clusters}
}
\value{
average cluster expression
}
\description{
calculate the average expression per cluster
}
