% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutlidataset_alignment.R
\name{load_additional_data}
\alias{load_additional_data}
\title{Load additional expression and annotation data}
\usage{
load_additional_data(
  data_name,
  data_file,
  data_version = NULL,
  data_taiga = TRUE,
  ann_name,
  ann_file,
  ann_version = NULL,
  ann_taiga = TRUE,
  data_type = ""
)
}
\arguments{
\item{data_name:}{if data_taiga = TRUE, then the data.name of the taiga file containing the expression data,
if data_taiga=FALSE, then the file path to the local folder containing the expression data. Assumes that genes
are labeled using ensembl IDs and that there are fewer samples than genes in the matrix, will transpose the matrix
so that rows are samples and columns are genes.}

\item{data_file:}{if data_taiga = TRUE, then the data.file of the taiga file containing the expression data,
if data_taiga = FALSE, then the name of the file of expression data}

\item{data_version:}{(optional) parameter to specify the version to pull from taiga for the expression data, default set to NULL}

\item{data_taiga:}{if TRUE then pulls the expression data from taiga, if FALSE then finds expression data in local folder}

\item{ann_name:}{if ann_taiga = TRUE, then the data.name of the taiga file containing the data annotations,
if ann_taiga=FALSE, then the file path to the local folder containing the annotations}

\item{ann_file:}{if ann_taiga = TRUE, then the data.file of the taiga file containing the data annotations,
if ann_taiga=FALSE, then the name of the file of data annotations}

\item{ann_version:}{(optional) parameter to specify the version to pull from taiga for the annotations, default set to NULL}

\item{ann_taiga:}{if TRUE (default) then pulls the annotations from taiga, if FALSE then finds cell line annotations in local folder}

\item{data_type:}{string added to the annotation file under the column type to specify the data, default is ""}
}
\value{
object containing expression matrix and annotations table
}
\description{
load additional expression and annotation files
}
