% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_helpers.R
\name{modified_mnnCorrect}
\alias{modified_mnnCorrect}
\title{MNN}
\usage{
modified_mnnCorrect(
  ref_mat,
  targ_mat,
  k1 = 20,
  k2 = 20,
  ndist = 3,
  subset_genes = NULL
)
}
\arguments{
\item{ref_mat:}{matrix of samples by genes of cPC corrected data that serves as the reference data in the MNN alignment.
In the standard Celligner pipeline this the cell line data.}

\item{targ_mat:}{matrix of samples by genes of cPC corrected data that is corrected in the MNN alignment and projected onto the reference data.
In the standard Celligner pipeline this the tumor data.}

\item{k1:}{the number of neighbors within the data being corrected (in standard pipeline the tumor data). By default this is 20.}

\item{k2:}{the number of neighbors within the reference data (in standard pipeline the cell line data). By default this is 20.}

\item{ndist:}{A numeric scalar specifying the threshold beyond which neighbors are to be ignored when computing correction vectors.
By default is 3.}

\item{subset_genes:}{the subset of genes used for identifying mutual nearest neighbors within the datasets. The set of differentially
expressed genes is usually passed here. By default is NULL, meaning all genes are used}
}
\value{
MNN object, containing the targ_mat corrected data and the mutual nearest neighbor pairs.
}
\description{
Mutual nearest neighbors correction. Modification of the scran::fastMNN (https://github.com/MarioniLab/scran).
Allows for separate k values per dataset, and simplifies some of the IO and doesn't use PCA reduction
}
