% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_methods.R
\name{run_Celligner}
\alias{run_Celligner}
\title{All methods to run Celligner and save the output, if desired}
\usage{
run_Celligner(
  cell_line_data_name = "public-20q4-a4b3",
  cell_line_data_file = "CCLE_expression_full",
  cell_line_version = NULL,
  cell_line_taiga = TRUE,
  cell_line_ann_name = "arxspan-cell-line-export-f808",
  cell_line_ann_file = "ACH",
  cell_line_ann_version = NULL,
  cell_line_ann_taiga = TRUE,
  tumor_data_name = "celligner-input-9827",
  tumor_data_file = "tumor_expression",
  tumor_version = NULL,
  tumor_taiga = TRUE,
  tumor_ann_name = "celligner-input-9827",
  tumor_ann_file = "tumor_annotations",
  tumor_ann_version = NULL,
  tumor_ann_taiga = TRUE,
  additional_annotations_name = "celligner-input-9827",
  additional_annotations_file = "CCLE_annotations",
  additional_annotations_version = NULL,
  additional_annotations_taiga = TRUE,
  hgnc_data_name = "hgnc-87ab",
  hgnc_data_file = "hgnc_complete_set",
  hgnc_version = NULL,
  hgnc_taiga = TRUE,
  save_output = NULL
)
}
\arguments{
\item{cell_line_data_name:}{if cell_line_taiga = TRUE, then the data.name of the taiga file containing the cell line expression data,
if cell_line_taiga=FALSE, then the file path to the local folder containing the cell line expression data}

\item{cell_line_data_file:}{if cell_line_taiga = TRUE, then the data.file of the taiga file containing the cell line expression data,
if cell_line_taiga=FALSE, then the name of the file of cell line expression data}

\item{cell_line_version:}{parameter to specify the version to pull from taiga for the cell line expression data, default set to NULL}

\item{cell_line_taiga:}{if TRUE then pulls the cell line expression data from taiga, if FALSE then finds cell line expression data in local folder}

\item{cell_line_ann_name:}{if cell_line_taiga = TRUE, then the data.name of the taiga file containing the cell line annotations,
if cell_line_taiga=FALSE, then the file path to the local folder containing the cell line annotations}

\item{cell_line_ann_file:}{if cell_line_taiga = TRUE, then the data.file of the taiga file containing the cell line annotations,
if cell_line_taiga=FALSE, then the name of the file of cell line annotations. If pulling from taiga, assumes that the file is the arxspan
file (could also use virtual release sample_info file), if not then assumes it is a local file containing the columns sampleID, lineage, subtype, and type=='CL'.}

\item{cell_line_ann_version:}{parameter to specify the version to pull from taiga for the cell line annotations, default set to NULL}

\item{cell_line_ann_taiga:}{if TRUE then pulls the cell line annotations from taiga, if FALSE then finds cell line annotations in local folder}

\item{tumor_data_name:}{if tumor_taiga = TRUE, then the data.name of the taiga file containing the tumor expression data,
if tumor_taiga=FALSE, then the file path to the local folder containing the tumor expression data.
If pulling from taiga, assumes that the file is the already create Celligner info file used in the Celligner manuscript,
if not then assumes it is a local file containing the columns sampleID, lineage, subtype, and type=='tumor'.}

\item{tumor_data_file:}{if tumor_taiga = TRUE, then the data.file of the taiga file containing the tumor expression data,
if tumor_taiga=FALSE, then the name of the file the tumor expression data}

\item{tumor_taiga:}{if TRUE then pulls the tumor expression data from taiga, if FALSE then finds tumor expression data in local folder}

\item{tumor_ann_name:}{if tumor_taiga = TRUE, then the data.name of the taiga file containing the tumor annotations,
if tumor_taiga=FALSE, then the file path to the local folder containing the tumor annotations}

\item{tumor_ann_file:}{if tumor_ann_taiga = TRUE, then the data.file of the taiga file containing the tumor annotations,
if tumor_ann_taiga=FALSE, then the name of the file the tumor annotations}

\item{tumor_version:}{parameter to specify the version to pull from taiga for the tumor annotations, default set to NULL}

\item{tumor_ann_taiga:}{if TRUE then pulls the tumor annotations from taiga, if FALSE then finds tumor annotations in local folder}

\item{additional_annotations_name:}{if additional_annotations_taiga = TRUE, then the data.name of the taiga file containing the additional annotations,
if additional_annotations_taiga=FALSE, then the file path to the local folder containing the additional annotations. Used to add more fine-grain subtype annotations
for the cell lines. If null, assumes there are no additional annotations.}

\item{additional_annotations_file:}{if additional_annotations_taiga = TRUE, then the data.file of the taiga file containing the additional annotations,
if additional_annotations_taiga=FALSE, then the name of the file the additional annotations. If null, assumes there are
no additional annotations.}

\item{additional_annotations_version:}{parameter to specify the version to pull from taiga for the tumor annotations, default set to NULL}

\item{additional_annotations_taiga:}{if TRUE then pulls the tumor annotations from taiga, if FALSE then finds tumor annotations in local folder}

\item{hgnc_data_name:}{if hgnc_taiga = TRUE, then the data.name of the taiga file containing the HGNC gene annotations,
if hgnc_taiga=FALSE, then the file path to the local folder containing the HGNC gene annotations}

\item{hgnc_data_file:}{if hgnc_taiga = TRUE, then the data.file of the taiga file containing the HGNC gene annotations,
if hgnc_taiga=FALSE, then the name of the file the HGNC gne annotations}

\item{hgnc_version:}{parameter to specify the version to pull from taiga for the HGNC gene annotations, default set to NULL}

\item{hgnc_taiga:}{if TRUE then pulls the HGNC gene annotations from taiga, if FALSE then finds HGNC gene annotations in local folder}

\item{save_output:}{by default is NULL and won't save output, to save output pass in a filepath of where to save the output}
}
\value{
seurat object of the Celligner-aligned data
}
\description{
run all parts of the Celligner pipeline
}
