% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_helpers.R
\name{run_cPCA_analysis}
\alias{run_cPCA_analysis}
\title{cPCA}
\usage{
run_cPCA_analysis(
  TCGA_dat,
  CCLE_dat,
  tumor_cluster_df,
  CL_cluster_df,
  pc_dims = NULL
)
}
\arguments{
\item{TCGA_dat:}{sample by genes matrix of scaled expression data}

\item{CCLE_dat:}{sample by genes matrix of scaled expression data}

\item{tumor_cluster_df:}{table of sample metadata that includes a column 'seurat_clusters',
containing transcriptional clusters in the TCGA data}

\item{CL_cluster_df:}{table of sample metadata that includes a column 'seurat_clusters',
containing transcriptional clusters in the CCLE data}

\item{pc_dims:}{numbers of cPCs calculated. If set to NULL (default) all cPCs will be calculated, if set to a value
then that number of cPCs will be approximated. Values input should be >= 4.}
}
\value{
contrastive principal component object containing cPC vectors and values
}
\description{
Run contrastive principal components analysis, first removing average cluster expression, to
}
