% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_helpers.R
\name{run_lm_stats_limma_group}
\alias{run_lm_stats_limma_group}
\title{Differentially expressed genes}
\usage{
run_lm_stats_limma_group(
  mat,
  phenos,
  covars = NULL,
  weights = NULL,
  target_type = "Gene",
  limma_trend = FALSE
)
}
\arguments{
\item{mat:}{Nxp data matrix of N cell lines and p genes}

\item{phenos:}{N vector of independent variables. Can be two-group labels as factors, bools, or can be numeric}

\item{covars:}{optional Nxk matrix of sample covariates}

\item{weights:}{optional N vector of precision weights for each data point}

\item{target_type:}{name of the column variable in the data (default 'Gene')}
}
\value{
table of gene level stata
}
\description{
Estimate linear-model stats for a matrix of data with respect to a group of phenotype variables
}
