% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_methods.R
\name{run_MNN}
\alias{run_MNN}
\title{Method to run mutual nearest neighbors batch correction}
\usage{
run_MNN(
  CCLE_cor,
  TCGA_cor,
  k1 = global$mnn_k_tumor,
  k2 = global$mnn_k_CL,
  ndist = global$mnn_ndist,
  subset_genes
)
}
\arguments{
\item{CCLE_cor:}{matrix of samples by genes of cPC corrected data that serves as the reference data in the MNN alignment.
In the default Celligner pipeline this the cell line data.}

\item{TCGA_cor:}{matrix of samples by genes of cPC corrected data that is corrected in the MNN alignment and projected onto the reference data.
In the default Celligner pipeline this the tumor data.}

\item{k1:}{the number of neighbors within the data being corrected (by default the tumor data). By default this
pulls from the global paramter mnn_k_tumor, which by default is 50.}

\item{k2:}{the number of neighbors within the reference data (by default the cell line data). By default this
pulls from the global parameter mnn_k_CL, which by default is 5.}

\item{ndist:}{A numeric scalar specifying the threshold beyond which neighbors are to be ingnored when computing correction vectors.
By default this pulls from the global parameter mnn_ndist, which by default is 3.}

\item{subset_genes:}{the subset of genes used for identifying mutual nearest neighbors within the datasets. The set of differentially
expressed genes is usually passed here.}
}
\value{
mutual nearest neighbors object with corrected data for the second dataset provided as input and the mutual nearest neighbors
}
\description{
run MNN batch correction to align data to a reference dataset
}
