% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutlidataset_alignment.R
\name{run_multidataset_alignment}
\alias{run_multidataset_alignment}
\title{All methods to run Celligner, with additional alignment of Met500 and PDX data, and save the output, if desired}
\usage{
run_multidataset_alignment(
  cell_line_data_name = "public-20q4-a4b3",
  cell_line_data_file = "CCLE_expression_full",
  cell_line_version = NULL,
  cell_line_taiga = TRUE,
  cell_line_ann_name = "arxspan-cell-line-export-f808",
  cell_line_ann_file = "ACH",
  cell_line_ann_version = NULL,
  cell_line_ann_taiga = TRUE,
  tumor_data_name = "celligner-input-9827",
  tumor_data_file = "tumor_expression",
  tumor_version = NULL,
  tumor_taiga = TRUE,
  tumor_ann_name = "celligner-input-9827",
  tumor_ann_file = "tumor_annotations",
  tumor_ann_version = NULL,
  tumor_ann_taiga = TRUE,
  additional_annotations_name = "celligner-input-9827",
  additional_annotations_file = "CCLE_annotations",
  additional_annotations_version = NULL,
  additional_annotations_taiga = TRUE,
  hgnc_data_name = "hgnc-87ab",
  hgnc_data_file = "hgnc_complete_set",
  hgnc_version = NULL,
  hgnc_taiga = TRUE,
  met500_data_name = "met500-fc3c",
  met500_data_file = "met500_TPM",
  met500_version = NULL,
  met500_taiga = TRUE,
  met500_ann_name = "met500-fc3c",
  met500_ann_file = "met500_ann",
  met500_ann_version = NULL,
  met500_ann_taiga = TRUE,
  Novartis_PDX_data_name = "pdx-data-3d29",
  Novartis_PDX_data_file = "Novartis_PDX_TPM",
  Novartis_PDX_version = NULL,
  Novartis_PDX_taiga = TRUE,
  Novartis_PDX_ann_name = "pdx-data-3d29",
  Novartis_PDX_ann_file = "Novartis_PDX_ann",
  Novartis_PDX_ann_version = NULL,
  Novartis_PDX_ann_taiga = TRUE,
  pediatric_PDX_data_name = "pdx-data-3d29",
  pediatric_PDX_data_file = "pediatric_PDX_TPM",
  pediatric_PDX_version = NULL,
  pediatric_PDX_taiga = TRUE,
  pediatric_PDX_ann_name = "pdx-data-3d29",
  pediatric_PDX_ann_file = "pediatric_PDX_ann",
  pediatric_PDX_ann_version = NULL,
  pediatric_PDX_ann_taiga = TRUE,
  save_output = NULL
)
}
\arguments{
\item{cell_line_data_name:}{if cell_line_taiga = TRUE, then the data.name of the taiga file containing the cell line expression data,
if cell_line_taiga=FALSE, then the file path to the local folder containing the cell line expression data}

\item{cell_line_data_file:}{if cell_line_taiga = TRUE, then the data.file of the taiga file containing the cell line expression data,
if cell_line_taiga=FALSE, then the name of the file of cell line expression data}

\item{cell_line_version:}{parameter to specify the version to pull from taiga for the cell line expression data, default set to NULL}

\item{cell_line_taiga:}{if TRUE then pulls the cell line expression data from taiga, if FALSE then finds cell line expression data in local folder}

\item{cell_line_ann_name:}{if cell_line_taiga = TRUE, then the data.name of the taiga file containing the cell line annotations,
if cell_line_taiga=FALSE, then the file path to the local folder containing the cell line annotations}

\item{cell_line_ann_file:}{if cell_line_taiga = TRUE, then the data.file of the taiga file containing the cell line annotations,
if cell_line_taiga=FALSE, then the name of the file of cell line annotations. If pulling from taiga, assumes that the file is the arxspan
file (could also use virtual release sample_info file), if not then assumes it is a local file containing the columns sampleID, lineage, subtype, and type=='CL'.}

\item{cell_line_ann_version:}{parameter to specify the version to pull from taiga for the cell line annotations, default set to NULL}

\item{cell_line_ann_taiga:}{if TRUE (default) then pulls the cell line annotations from taiga, if FALSE then finds cell line annotations in local folder}

\item{tumor_data_name:}{if tumor_taiga = TRUE, then the data.name of the taiga file containing the tumor expression data,
if tumor_taiga=FALSE, then the file path to the local folder containing the tumor expression data.}

\item{tumor_data_file:}{if tumor_taiga = TRUE, then the data.file of the taiga file containing the tumor expression data,
if tumor_taiga=FALSE, then the name of the file the tumor expression data}

\item{tumor_version:}{parameter to specify the version to pull from taiga for the tumor expression data, default set to NULL}

\item{tumor_taiga:}{if TRUE (default) then pulls the tumor expression data from taiga, if FALSE then finds tumor expression data in local folder}

\item{tumor_ann_name:}{if tumor_taiga = TRUE, then the data.name of the taiga file containing the tumor annotations,
if tumor_taiga=FALSE, then the file path to the local folder containing the tumor annotations}

\item{tumor_ann_file:}{if tumor_ann_taiga = TRUE, then the data.file of the taiga file containing the tumor annotations,
if tumor_ann_taiga=FALSE, then the name of the file the tumor annotations. If pulling from taiga, assumes that the file is the already create Celligner info file used in the Celligner manuscript,
if not then assumes it is a local file containing the columns sampleID, lineage, subtype, and type=='tumor'.}

\item{tumor_ann_version:}{parameter to specify the version to pull from taiga for the tumor annotations, default set to NULL}

\item{tumor_ann_taiga:}{if TRUE then pulls the tumor annotations from taiga, if FALSE then finds tumor annotations in local folder}

\item{additional_annotations_name:}{if additional_annotations_taiga = TRUE, then the data.name of the taiga file containing the additional annotations,
if additional_annotations_taiga=FALSE, then the file path to the local folder containing the additional annotations. Used to add more fine-grain subtype annotations
for the cell lines. If null, assumes there are no additional annotations.}

\item{additional_annotations_file:}{if additional_annotations_taiga = TRUE, then the data.file of the taiga file containing the additional annotations,
if additional_annotations_taiga=FALSE, then the name of the file the additional annotations. If null, assumes there are
no additional annotations.}

\item{additional_annotations_version:}{parameter to specify the version to pull from taiga for the tumor annotations, default set to NULL}

\item{additional_annotations_taiga:}{if TRUE then pulls the tumor annotations from taiga, if FALSE then finds tumor annotations in local folder}

\item{hgnc_data_name:}{if hgnc_taiga = TRUE, then the data.name of the taiga file containing the HGNC gene annotations,
if hgnc_taiga=FALSE, then the file path to the local folder containing the HGNC gene annotations}

\item{hgnc_data_file:}{if hgnc_taiga = TRUE, then the data.file of the taiga file containing the HGNC gene annotations,
if hgnc_taiga=FALSE, then the name of the file the HGNC gne annotations}

\item{hgnc_version:}{parameter to specify the version to pull from taiga for the HGNC gene annotations, default set to NULL}

\item{hgnc_taiga:}{if TRUE then pulls the HGNC gene annotations from taiga, if FALSE then finds HGNC gene annotations in local folder}

\item{met500_data_name:}{Met500 expression, default pulls from taiga, this the data_name of the taiga dataset, or path to folder if using met500_taiga=F}

\item{met500_data_file:}{default pulls from taiga, this the data_file of the taiga dataset, or name of local file if using met500_taiga=F}

\item{met500_version:}{default NULL, used to specify version of taiga dataset}

\item{met500_taiga:}{if TRUE (default) pulls Met500 expression from taiga dataset, if FALSE reads from local}

\item{met500_ann_name:}{Met500 annotations, default pulls from taiga, this the data_name of the taiga dataset, or path to folder is using met500_ann_taiga=F}

\item{met500_ann_file:}{Met500 annotations, default pulls from taiga, this the data_file of the taiga dataset, or name of local file is using met500_ann_taiga=F}

\item{met500_ann_version:}{default NULL, used to specify version of taiga dataset}

\item{met500_ann_taiga:}{if TRUE (default) pulls met500 annotations from taiga dataset, if FALSE reads from local}

\item{Novartis_PDX_data_name:}{Novartis PDX expression, default pulls from taiga, this the data_name of the taiga dataset, or path to folder if using Novartis_PDX_taiga=F}

\item{Novartis_PDX_data_file:}{default pulls from taiga, this the data_file of the taiga dataset, or name of local file if using Novartis_PDX_taiga=F}

\item{Novartis_PDX_version:}{default NULL, used to specify version of taiga dataset}

\item{Novartis_PDX_taiga:}{if TRUE (default) pulls Novartis PDX expression from taiga dataset, if FALSE reads from local}

\item{Novartis_PDX_ann_name:}{Novartis PDX annotations, default pulls from taiga, this the data_file of the taiga dataset, or path to folder is using met500_ann_taiga=F}

\item{Novartis_PDX_ann_file:}{Novartis PDX annotations, default pulls from taiga, this the data_name of the taiga dataset, or name of local file is using Novartis_PDX_ann_taiga=F}

\item{Novartis_PDX_ann_version:}{default NULL, used to specify version of taiga dataset}

\item{Novartis_PDX_ann_taiga:}{if TRUE (default) pulls Novartis PDX annotations from taiga dataset, if FALSE reads from local}

\item{pediatric_PDX_data_name:}{pediatric PDX expression, default pulls from taiga, this the data_name of the taiga dataset, or path to folder if using pediatric_PDX_taiga=F}

\item{pediatric_PDX_data_file:}{default pulls from taiga, this the data_file of the taiga dataset, or name of local file if using pediatric_PDX_taiga=F}

\item{pediatric_PDX_version:}{default NULL, used to specify version of taiga dataset}

\item{pediatric_PDX_taiga:}{if TRUE (default) pulls pediatric PDX expression from taiga dataset, if FALSE reads from local}

\item{pediatric_PDX_ann_name:}{Pediatric PDX annotations, default pulls from taiga, this the data_name of the taiga dataset, or path to folder is using pediatric_PDX_ann_taiga=F}

\item{pediatric_PDX_ann_file:}{Pediatric PDX annotations, default pulls from taiga, this the data_file of the taiga dataset, or name of local file is using pediatric_PDX_ann_taiga=F}

\item{pediatric_PDX_ann_version:}{default NULL, used to specify version of taiga dataset}

\item{pediatric_PDX_ann_taiga:}{if TRUE (default) pulls pediatric PDX annotations from taiga dataset, if FALSE reads from local}

\item{save_output:}{by default is NULL and won't save output, to save output pass in a filepath of where to save the output}
}
\value{
seurat object of the Celligner-aligned data
}
\description{
run all parts of the Celligner pipeline, with alignment of additional datasets
}
